Remove TaskSolver, NotificationReceiver modules
