+++ HW04/ekaterina_balabanovich/lib/member.rb
@@ -1,17 +1,28 @@
 # frozen_string_literal: true
 
-require_relative 'task_solver'
-require_relative 'notification_receiver'
-
 class Member
-  include TaskSolver
-  include NotificationReceiver
-  attr_reader :name, :surname
+  attr_reader :name, :surname, :notifications, :homeworks
 
   def initialize(name, surname)
     @name = name
     @surname = surname
     @homeworks = []
     @notifications = []
+  end
+
+  def get_notifications(notification)
+    @notifications << notification
+  end
+
+  def mark_as_read!
+    @notifications.clear
+  end
+
+  def get_homework(homework)
+    @homeworks << homework
+  end
+
+  def delete_homework(homework)
+    @homeworks.delete(homework)
   end
 end
